Temporal reunion of stale changes.
This commit aggregates a number of unrelated changesets accumulating on
this local machine. No sane commit message is feasible, as much of the
original impetus for these changes has been carelessly forgotten.

That said, the most crucial change is probably the elimination of the
"update.bash" shell script -- which, frankly, never behaved as expected
and could be expected to fail in common edge cases (e.g., dependency
revisions in updated application versions). To safely update this
application, users are instead advised to simply rerun the
platform-specific installer of their choice.