Simulator progress status x 16.
This is the last in a commit chain displaying human-readable progress
status emitted by BETSE during every simulation phase as the text of a
currently hidden simulator text widget. Notably, this commit hallmarks
the end of the 0.8.* development cycle for both BETSE and BETSEE. A
feature freeze is now in effect, paving the way for the upcoming stable
releases of BETSE and BETSEE 0.9.0. (Irregular singularities agglutinate!)