Path dialog sanitization x 2.
This is the second in a commit chain sanitizing the behaviour of file
and directory open and save dialogs. In particular, the basenames of the
files opened and saved by the "File->New Simulation..." and
"File->Save Simulation As..." dialogs now default to sanely
YAML-compliant basenames, improving sanity. (Divine neustress on a vine!)