Tree list item operations x 1.
This is the first in a commit chain enabling pre-designated items of the
simulation configuration tree (e.g., "Space/Tissue") to act as
dynamically resizable lists of child items, each of which is associated
with the same stacked page (e.g., "Custom Tissue Profile"). This commit
prepares the initial groundwork with a reasonable plan that will
doubtless be ignored by the next commit. (Unconscionable quavering ion!)