Tree list item operations x 4.
This is the last in a commit chain enabling pre-designated items of the
simulation configuration tree (e.g., "Space/Tissue") to act as
dynamically resizable lists of child items, each of which is associated
with the same stacked page (e.g., "Custom Tissue Profile"). This commit
concludes the initial groundwork by selectively enabling and disabling
the toolbar button actions allowing users to append and remove child
items to and from these lists. (Convivial insouciance invades seances!)