Custom tissue profile switching x 5.
This is the next in a commit chain permitting end users to configure
custom tissue profiles. Specifically, this commit finalizes the
generalization of the "QBetseeSimConfTissueStackedWidgetPagerABC"
superclass to support both default and custom tissue profiles. Since
this mostly addresses this commit chain, the next commit is intended to
be the last in this chain. (Superfluidic apostasy of the ecstatic ion!)