Custom tissue profile switching x 8.
This is the next in a commit chain permitting end users to configure
custom tissue profiles. Specifically, this commit:

* Permits application-specific Qt objects to be reinitialized in a
  general-purpose manner via the new optional "is_reinitable" parameter
  passed to the existing QBetseeObjectMixin.init() method, naturally
  defaulting to "False".
* Reinitializes the custom tissue profile stacked widget page via the
  aforementioned parameter immediately before displaying that page.

Since doing so suffices to finalize this commit chain, the next commit
will inaugurate a new commit chain permitting new custom tissue profiles
to be added and existing custom tissue profiles to be removed. (Ineluctable pluck!)