Custom tissue profile operations x 2.
This is the first in a commit chain permitting end users to append new
custom tissue profiles to and remove existing custom tissue profiles
from the list of these profiles for the current open simulation
configuration. Specifically, this commit implements most of the body of
the QBetseeSimConfTreeWidget._append_tree_item() method, now generalized
to support creation of any arbitrary tree items. (Methodical threnody!)