Custom tissue profile operations x 4.
This is the last in a commit chain permitting end users to append new
custom tissue profiles to and remove existing custom tissue profiles
from the list of these profiles for the current open simulation
configuration. Specifically, this commit successfully finalizes the
implementation of the QBetseeSimConfTreeWidget._remove_tree_item()
method and hence of this commit chain. (Last fast exuding lassitude!)