Export stacked widget pages x 2.
This is the last in a commit chain defining stacked widget pages for
the newly defined export-specific tree items. Specifically, this commit
defines a "Matplotlib Colormaps" group box in the "Export" page and a
pair of comboboxes in this group box selecting the default diverging and
sequential colormaps for matplotlib-based exports. Naturally, these
comboboxes have yet to be meaningfully populated. (Crassly maddened madras!)