Export stacked widget pages x 14.
This is the next in a commit chain defining stacked widget pages for
the newly defined export-specific tree items. Specifically, this commit
completes the implementation of the
QBetseeSimConfTreeWidget._init_items() method -- which now correctly
informs the top-level stacked widget of mappings between items of the
top-level tree widget and page widgets of this stacked widget. (Toboggan noggin!)